# pkg/api/openapi.yaml
openapi: 3.1.0
info:
  title: OffGrid Local API
  version: 1.0.0
  description: Stable contract shared by the OffGrid service, web UI, desktop app, CLI, and external agents.
servers:
  - url: http://127.0.0.1:11611
security:
  - sessionCookie: []
  - bearerAuth: []
paths:
...
  /v1/sessions/{name}/turn/events:
    parameters:
      - $ref: '#/components/parameters/SessionName'
      - { name: id, in: query, required: true, schema: { type: string } }
    get:
      operationId: followSessionTurn
      description: Authorized snapshot recovery. Replays the entire current partial output, followed by live deltas. A new attachment must replace its previous partial output, not append a replay. Slow/disconnected readers do not control execution.
      responses:
        '200':
          description: Session phase/delta/done/error SSE stream
          content:
            text/event-stream:
              schema: { type: string }
        '409': { $ref: '#/components/responses/Error' }
...
  /api/v2/jobs:
    get:
      operationId: listDurableJobs
      responses:
        '200':
          description: Owner-scoped task history, including linked child tasks
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/AgentTask' } }
    post:
      operationId: submitDurableTask
      description: Saves an owner-scoped task before execution. Identical actor/request ID retries return the original task; conflicting reuse returns 409. Computer access is requested as a durable interruption, not a submission prerequisite.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [prompt, model, request_id]
              properties:
                prompt: { type: string, minLength: 1, maxLength: 32000 }
                model: { type: string }
                request_id: { type: string, minLength: 8, maxLength: 128 }
                style: { type: string, enum: [react, cot, plan-execute] }
                max_iterations: { type: integer, minimum: 1, maximum: 50 }
      responses:
        '202':
          description: Persisted task
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AgentRunResponse' }
        '400': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/Error' }
...
  /api/v2/jobs/{id}:
    delete:
      operationId: deleteDurableJob
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Terminal history removed; referenced children remain protected until the parent is removed
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
    get:
      operationId: getDurableJobSnapshot
      description: Reads the existing owner-scoped agent task, including the latest durable event cursor. No separate task history is created.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Current persisted job snapshot
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AgentRunResponse' }
        '404': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/Error' }
...
  /api/v2/jobs/{id}/artifact:
    get:
      operationId: downloadTaskArtifact
      description: Checks current task ownership, a committed artifact reference, size and SHA256 before delivering an attachment. Digest knowledge alone does not grant access.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: digest, in: query, required: true, schema: { type: string, pattern: '^[a-f0-9]{64}$' } }
      responses:
        '200':
          description: Verified artifact bytes
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        '404': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/Error' }
  /api/v2/jobs/{id}/{action}:
    post:
      operationId: commandDurableJob
      description: All clients share the durable runner. Pause settles or labels in-flight effects uncertain. Takeover revokes the host session. Steering invalidates unexecuted proposals, is idempotent by request_id, and requires explicit resume. Reconnect requests fresh local consent, never an automatic retry.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: action, in: path, required: true, schema: { type: string, enum: [approve, deny, cancel, pause, resume, takeover, steer, reconnect, reconcile] } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                approval_id: { type: string }
                call_id: { type: string }
                result: { type: string, maxLength: 16000 }
                request_id: { type: string, minLength: 8, maxLength: 128 }
                instruction: { type: string, maxLength: 16000 }
                mode: { type: string, enum: [app, browser] }
                target: { type: string, maxLength: 256 }
                url: { type: string, maxLength: 2048 }
                async: { type: boolean }
      responses:
        '202':
          description: Persisted task state
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AgentRunResponse' }
        '400': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/Error' }
  /api/v2/jobs/{id}/events:
    get:
      operationId: replayDurableJobEvents
      description: Owner-only ordered activity metadata and current snapshots. Last-Event-ID resumes from a persisted sequence. An expired or ahead cursor receives an explicit snapshot_recovery event with the current snapshot and event_cursor. Snapshot recovery never repeats execution. Disconnect does not cancel a job. Event history retains the last 256 metadata entries; completed snapshots remain.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: Last-Event-ID, in: header, schema: { type: string, pattern: '^[0-9]{1,19}$' } }
      responses:
        '200':
          description: SSE activity events contain TaskActivity; snapshot and snapshot_recovery events contain AgentRunResponse plus type. Event IDs are decimal sequence strings. Terminal snapshots close the stream.
          content:
            text/event-stream:
              schema: { type: string }
        '400': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/Error' }
  /v1/agents/tasks/{id}/events:
    get:
      operationId: streamAgentProgress
      summary: Stream durable agent snapshots without owning execution
      description: Starts with the latest saved snapshot, then sends changed snapshots and heartbeat comments. Reconnect to recover the latest bounded preview; this is not an event-history replay endpoint. Disconnect does not cancel work. Only the run owner can subscribe.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: SSE data frames contain AgentRunResponse plus type=status, done, approval_required or error. A preview is provisional, never a completed answer. Terminal snapshots close the stream.
          content:
            text/event-stream:
              schema: { type: string }
        '404': { description: Run not found or not owned by the current actor }
        '503': { description: Agent storage unavailable }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    SessionName:
      in: path
      name: name
      required: true
      schema: { type: string, minLength: 1, maxLength: 128 }
  responses:
    Error:
      description: API error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          oneOf:
            - type: string
            - type: object
              required: [message]
              properties:
                message: { type: string }
                type: { type: string }
                code: { type: string }
...
    ToolApproval:
      type: object
      required: [id, run_id, call_id, actor, tool, arguments, canonical_arguments, expires_at]
      properties:
        id: { type: string }
        run_id: { type: string }
        call_id: { type: string }
        actor: { type: string }
        tool: { type: string }
        arguments: { type: object, additionalProperties: true }
        canonical_arguments: { type: string, description: Exact canonical JSON for review without JavaScript numeric precision loss. }
        capability: { type: object, additionalProperties: true }
        expires_at: { type: string, format: date-time }
...
    AgentStep:
      type: object
      properties:
        id: { type: integer }
        type: { type: string }
        content: { type: string }
        tool_name: { type: string }
        tool_args: { type: string }
        tool_result: { type: string }
        authorization: { type: string, description: Durable record of exact approval or the immutable automatic session policy used for this action. }
        timestamp: { type: string, format: date-time }
    AgentProgress:
      type: object
      required: [phase, iteration, preview, truncated, started_at, updated_at]
      properties:
        phase: { type: string, enum: [queued, loading, processing, generating, preparing_tool, tool, approval] }
        iteration: { type: integer }
        tool: { type: string }
        preview: { type: string, description: Bounded public model response preview, separate from completed output and model context. }
        truncated: { type: boolean }
        started_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    AgentInputRequest:
      type: object
      required: [id, call_id, kind, mode, target]
      properties:
        id: { type: string }
        call_id: { type: string }
        kind: { type: string, enum: [computer] }
        mode: { type: string, enum: [app, browser] }
        target: { type: string, description: Model-proposed label only; never a host identity or consent grant. }
        url: { type: string, description: Proposed HTTPS address for local review. }
        previous_session: { type: string, description: Revoked session identity; never reusable authority. }
    TaskPlanItem:
      type: object
      required: [id, title, state]
      properties:
        id: { type: string }
        title: { type: string }
        state: { type: string, enum: [pending, active, done] }
        evidence_steps: { type: array, items: { type: integer } }
    TaskChild:
      type: object
      required: [id, spec]
      properties:
        id: { type: string }
        spec:
          type: object
          required: [key, goal, depends_on, tools]
          properties:
            key: { type: string }
            goal: { type: string }
            depends_on: { type: array, items: { type: string } }
            tools: { type: array, items: { type: string } }
    TaskContext:
      type: object
      required: [window, estimated_tokens, compactions, archived_messages]
      properties:
        window: { type: integer }
        estimated_tokens: { type: integer, description: Conservative byte-based estimate, not tokenizer output }
        measured_prompt_tokens: { type: integer, description: Last runtime-reported prompt usage }
        compactions: { type: integer }
        archived_messages: { type: integer }
    AgentRunResponse:
      type: object
      required: [output, task_id, run_id, steps, status, pending_approval, resumable]
      properties:
        prompt: { type: string, description: Included in the owner-scoped job GET only, not the event stream. }
        model: { type: string }
        error_code: { type: string }
        artifacts:
          type: array
          items:
            type: object
            required: [name, format, sha256, bytes, verified, check]
            properties:
              name: { type: string }
              format: { type: string, enum: [text, markdown, json, csv] }
              sha256: { type: string }
              bytes: { type: integer }
              rows: { type: integer }
              columns: { type: integer }
              verified: { type: boolean }
              check: { type: string, enum: [stored_bytes_sha256_and_format] }
        parent_id: { type: string }
        children: { type: [array, 'null'], items: { $ref: '#/components/schemas/TaskChild' } }
        can_steer: { type: boolean, description: Saved state permits a follow-up instruction. Completed child history does not block steering; active delegation and uncertain effects do. Rechecked at execution. }
        plan: { type: [array, 'null'], items: { $ref: '#/components/schemas/TaskPlanItem' } }
        context:
          anyOf:
            - { $ref: '#/components/schemas/TaskContext' }
            - { type: 'null' }
        last_access:
          anyOf:
            - { $ref: '#/components/schemas/AgentInputRequest' }
            - { type: 'null' }
        instructions:
          type: [array, 'null']
          items:
            type: object
            required: [request_id, text, at]
            properties:
              request_id: { type: string }
              text: { type: string }
              at: { type: string, format: date-time }
        pending_input:
          anyOf:
            - { $ref: '#/components/schemas/AgentInputRequest' }
            - { type: 'null' }
        event_cursor: { type: string, pattern: '^[0-9]{1,19}$', description: Persisted activity cursor. Only advance it after applying the snapshot. }
        computer_session: { type: string }
        computer_approval_mode: { type: string, enum: [ask_every_time, scoped_changes, full_task] }
        computer_expected_text: { type: string }
        computer_session_expired: { type: boolean, description: Restart invalidated local consent. Historical work may be inspected or reconciled, but cannot resume against the old target. }
        output: { type: string }
        task_id: { type: string }
        run_id: { type: string }
        status: { type: string, enum: [pending, running, waiting_for_approval, waiting_for_input, waiting_for_children, interrupted, uncertain, completed, failed, cancelled] }
        error: { type: string }
        pending_approval:
          anyOf:
            - { $ref: '#/components/schemas/ToolApproval' }
            - { type: 'null' }
        resumable: { type: boolean, description: Whether this run has a safe checkpoint for explicit resume. Legacy history without checkpoints is read-only. }
        started_at: { type: [string, 'null'], format: date-time }
        progress:
          anyOf:
            - { $ref: '#/components/schemas/AgentProgress' }
            - { type: 'null' }
        uncertain_call_id: { type: string }
        uncertain_call:
          type: object
          required: [tool, arguments]
          properties:
            tool: { type: string }
            arguments: { type: object, additionalProperties: true }
        steps: { type: array, items: { $ref: '#/components/schemas/AgentStep' } }
    AgentTask:
      type: object
      required: [id, prompt, status, created_at]
      properties:
        deletable: { type: boolean, description: Whether this actor may remove this terminal run or interrupted record without recovery state; the server rechecks ownership and execution state at deletion time. }
        id: { type: string }
        parent_id: { type: string }
        prompt: { type: string }
        status: { type: string }
        model: { type: string }
        actor: { type: string }
        pending_approval: { $ref: '#/components/schemas/ToolApproval' }
        result: { type: string }
        error: { type: string }
        error_code: { type: string, description: Stable machine-readable failure code for recovery and automation. }
        steps: { type: array, items: { $ref: '#/components/schemas/AgentStep' } }
        created_at: { type: string, format: date-time }
        started_at: { type: string, format: date-time }
        completed_at: { type: string, format: date-time }
...